Move error directory up a level
