+++ npm-packages/meteor-vite/src/bin/debug/watch-mocks.ts
@@ -1,6 +1,6 @@
 import { Check, TestLazy, TsModules } from '../../../test/__mocks';
 import MeteorPackage from '../../meteor/package/components/MeteorPackage';
-import { MeteorViteError } from '../../vite/error/MeteorViteError';
+import { MeteorViteError } from '../../error/MeteorViteError';
 import ViteServer from './vite-server';
 
 /**
