+++ npm-packages/meteor-vite/src/meteor/package/components/ModuleExport.ts
@@ -1,4 +1,4 @@
-import { ErrorMetadata, MeteorViteError } from '../../../vite/error/MeteorViteError';
+import { ErrorMetadata, MeteorViteError } from '../../../error/MeteorViteError';
 import { PackageSubmodule } from './PackageSubmodule';
 import { ModuleExportData } from '../../parser/Parser';
 import { METEOR_STUB_KEY } from '../StubTemplate';
